requirements.txt
aws_cdk.aws_ec2
aws_cdk.aws_efs
netaddr